Display user-friendly key usage descriptions for certificate purpose
When displaying the certificate purpose to the user, include
human-readable descriptions of both the "key usage" and "extended key
usage" fields.

This change includes a built-in set of common extended key usage OIDs.
But OIDs are organized into a hierarchical registry that organizations
can extend.  To support organization-specific extensions, a method is
provided for app developers to register additional OIDs that weren't
compiled-in.