Remember user's alias selection (including Cancel) per host:port
This change improves the caching of the user's alias selections.

1) If a single process connects to more than one server, different
certificate aliases can be used to connect to different servers.

2) If the user presses Cancel on the alias selection dialog, that choice
is now remembered. Previously, the user's "null" choice was treated the
same as "no cached value yet," resulting in the display of additional
dialogs when chooseClientAlias was called again. This was a problem
because some versions of Java call chooseClientAlias repeatedly for a
single SSL connection attempt.