Show cert selections with user-friendly output; allow app customization
When we ask the user to select from a list of certificates, use an
object of type IdentityKeyCertFormatter to format the certificate
choices for display. Install a user-friendly formatter by default.

Application developers can install a different/custom formatter as
needed, via the following idiom:

    AbstractSelectorKeyManager
        .configureSwingKeyManagerAsDefault(config)
        .getIdentityKeyChooser()
        .setCertFormatter(DebugCertFormatter.INSTANCE);

An app-specific formatter object can be provided to the setCertFormatter
method to customize the display of certificates.