Align chooser dialog titles and icons with the parent application
Allow application developers to register the name of their application,
then use that in the titles of the windows we display to the user.

Allow application developers to register the parent window for the CAC
dialog. Center the windows we display over that parent, and copy its
icons onto our windows.