Show in task list when dialogs open. Also, show icon in system tray. When system tray icon clicked bring dialog prompting to front.
Under some circumstances the dialog
could be covered by other windows
and hard to find, or rarely, the dialog
would be completely hidden.

Tested under Windows 10 and Ubuntu 20.04.
Windows supports both the task list and
system tray features.
Ubuntu no longer has a system tray,
so that feature does not work on Ubuntu.
Also, the title on Ubuntu shows up as
"java-lang-Thread". However, it is still
useful in helping to find the dialog
if it is hidden.