Fix hang when using opensc.
Root issue: disposing defaultParent before
busyDialog appeared. `defaultParent`
is the parent for busyDialog.

Added missing @Override annotations.

Made sure every method that does
Swing operations does them from the
"Swing thread".

Replaced the "pre-create" of the busy dialog
with, just creating it after the timer goes off.

Reduced the time for the busy dialog to
appear from 3 seconds to 1 second.

Made it safe to call hideBusy if busy ref is null.
That can happen if the operation completes
before we need to show the busy dialog.

Created a helper method, invokeAndWait
that can return a result from the action
it waited to complete.